Break judgment_client_test.py into separate test files
